apps/ai-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.59.3
pinecone-client==5.0.1
firebase-admin==6.6.0
python-dotenv==1.0.1
sse-starlette==2.1.3
httpx==0.28.1
pydantic-settings==2.7.1
numpy==2.2.1
